+++ lib/features/ai/model/ai_config.dart
@@ -8,6 +8,7 @@ enum InferenceProviderType {
   anthropic,
   gemini,
   genericOpenAi,
+  mistral,
   nebiusAiStudio,
   openAi,
   openRouter,